requirements.txt
absl-py==1.4.0
aiohttp==3.8.5
aiosignal==1.3.1
amqp==5.1.1
appnope==0.1.3
asgiref==3.7.2
asttokens==2.4.0
async-timeout==4.0.3
attrs==22.2.0
backcall==0.2.0
beartype==0.15.0
beautifulsoup4==4.12.2
billiard==4.1.0
binance-futures-connector==4.0.0
bitrix24-rest==1.1.1
bs4==0.0.1
celery==5.3.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
comm==0.1.4
contourpy==1.0.7
cycler==0.11.0
debugpy==1.8.0
decorator==5.1.1
Django==4.2.4
django-cors-headers==4.2.0
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
drf-yasg==1.21.7
exceptiongroup==1.1.3
executing==1.2.0
fast-bitrix24==1.5.16
flatbuffers==23.3.3
fonttools==4.39.3
frozenlist==1.4.0
h11==0.14.0
icontract==2.6.2
idna==3.4
inflection==0.5.1
ipykernel==6.25.2
ipython==8.15.0
jedi==0.19.0
jupyter_client==8.3.1
jupyter_core==5.3.1
kiwisolver==1.4.4
kombu==5.3.1
llvmlite==0.41.0
matplotlib==3.7.1
matplotlib-inline==0.1.6
mediapipe-silicon==0.9.1
more-itertools==10.1.0
multidict==6.0.4
nest-asyncio==1.5.8
numba==0.58.0
numpy==1.24.2
opencv-contrib-python==4.7.0.72
outcome==1.2.0
packaging==23.1
pandas==2.1.1
parso==0.8.3
pexpect==4.8.0
phonenumbers==8.13.19
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.10.0
prompt-toolkit==3.0.39
protobuf==3.20.3
psutil==5.9.5
psycopg2-binary==2.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pycryptodome==3.19.0
Pygments==2.16.1
pyparsing==3.0.9
PyQt3D==5.15.5
PyQt5==5.15.7
PyQt5-sip==12.11.0
PyQt6==6.5.2
PyQt6-3D==6.5.0
PyQt6-Charts==6.5.0
PyQt6-DataVisualization==6.5.0
PyQt6-NetworkAuth==6.5.0
PyQt6-sip==13.5.2
PyQt6-WebEngine==6.5.0
PyQtChart==5.15.6
PyQtDataVisualization==5.15.5
PyQtNetworkAuth==5.15.5
PyQtPurchasing==5.15.5
PyQtWebEngine==5.15.6
PySocks==1.7.1
pyTelegramBotAPI==4.14.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
pyzmq==25.1.1
redis==5.0.0
requests==2.28.1
selenium==4.13.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.2
TBB==0.2
telebot==0.0.5
tornado==6.3.3
tqdm==4.64.1
traitlets==5.10.1
trio==0.22.2
trio-websocket==0.10.4
typing_extensions==4.7.1
tzdata==2023.3
undetected-chromedriver==3.5.3
uritemplate==4.1.1
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.6
websocket-client==1.6.4
websockets==11.0.3
wsproto==1.2.0
yarl==1.9.2